Consolidate midplane logic into midplane_refinement.py
Move register_centroids_to_fsavg, refine_midline_lr_shift, and all
supporting helpers (_ASEG_LR_PAIRS, _prepare_lr_pair_labels,
_score_midline_shift) from fastsurfer_cc.py into midplane_refinement.py.

Add MidplaneTransformResult dataclass and find_midplane_transform(),
a single entry point that dispatches across all four midplane methods
("center", "fsaverage", "fsaverage_symmetry", "fsaverage_distance_map")
and returns a uniform result.

Replace the scattered midplane dispatch block in main() with one call
to find_midplane_transform(). Remove the five debug-output CLI flags
(--midplane_distance_map, --midplane_fit_mask, etc.) and their
corresponding save blocks, which were used only during development.
